DB2 migration support
DB2 will not allow you to rename a table if it's got a
constraint on it (a unique or foreign key constraint).

This fix changes the migrations so that the unique or FK
constraints are dropped from tables before renaming and then
restoring the unique FK constraints. This works for DB2 and
other DBMSs that support FK constraints such as MySQL with
InnoDB and PostgreSQL.

Also, for DB2, give a name to the unique constraints so that
they can be manipulated.

Fixes bug 1188785

Change-Id: I7cf6ab42084e43d827ed827c64025e61e72a4672