Commit transaction in migration
Some databases (e.g., DB2) would deadlock on migration 16 because a
lock was held on the domain table for an uncommitted alter and then
SQLAlchemy queried the domain table info in a separate transaction.

The fix is to commit the ALTER statements to release the lock so
that the domain table schema query will not block.

I've run this with sqlite, mysql, and postgresql and they all worked.

Part of fix for bug 1188785

Change-Id: Ic540a6cb09a0c525df7aaea55b64af96f0dd87c7