Correct the resolving api logic in stat middleware
If the admin_port string is a substring of public_port string, all the
requests would be stated as 'admin' either their real dest port is
admin_port or public_port. It's due to the incorrect logic in stat
middleware.
For example, if public_port = 35000, admin_port = 5000, the first
judgement branch: "if str(CONF.admin_port) in host" in
StatsMiddleware::_resolve_api() would always return "True" either the
host port number equal to 5000 or 35000, so that the following judgement
branches would be incorrectly ignored.

Fixes bug #1189121
Change-Id: I1086b7d11f83dd218d66376f79747a1f720eb807