Move identity ldap backend from directory to file
The LDAP  backend for identity was in a subdirectory as opposed to a file
due to a clash between the name of the local package and the standard python package
both being called ldap.
This change makes the ldap backend look like the other backend.
The mechanism that fixes it is:

from __future__ import absolute_import

Change-Id: Id9ee518238ae22e5128351094eb3764a831cc083