Deprecate kvs token backend
This backend is not usable in any production environment. All OpenStack
environments will already have a SQL DB, and if someone does not want to
use the DB they can use the memcache backend.

Fixes bug 1188301 and bug 1188370

DocImpact This backend should not be mentioned in documentation, as it
is not production grade and is deprecated.

Change-Id: I41b147bcc70b79b4fc6df50b242a73cfcad33114