Pluggable Remote User
Select the code to handle REMOTE_USER based on a config file option

Fixes the REMOTE_USER logic to get the domain name from
REALM, which is the least surprise option.

Disregards the auth_data passed in, as we should be using REMOTE_USER
to get the user name.

External Plugin is now executed in conjunction with the auth methods,
as opposed to in place of them.

DocImpact

blueprint pluggable-remote-user

Change-Id: I9dda6dbe073f03806bdf539db6faa01644109f1c