Clean up use of token_provider manager in tests
With 19081b834991d263d84c761dcf422a8c9faf40a1, the token_provider
is created in load_backends() so it doesn't have to be created by
the tests.

This makes the tests using token_provider_api consistent with how
other backends are used in the tests.

This change only affects unit tests.

Change-Id: I9c05c7876bd14fb28e77fe1e49dd509a2a1cb788