Initial implementation of unified-logging
This is the base implementation of a unified logging solution for
Keystone from Oslo-incubator. More work is still needed to refactor the
rest of Keystone such that it is completely dependent on the
implementation from Oslo and not the older keystone/common/logging.py
implementation, this is also noted in keystone/common/config.py.

blueprint unified-logging-in-keystone

Change-Id: I711cbac8edd887c52114fb13327e37124ea86737