Implement list limiting support in driver backends
Provides an optional limit to the number of rows that will be
returned by a backend from a list_{entity} call.  The limit is specified
in the configuration file, and allows for an overall general limit as
well as an individual limit for a given driver.  By default, there is
no limit.

Limitations:
    - The list limit is not yet handled by LDAP, rather this remains
      implemented in the final wrap collection - a subsequent patch
      will provide the support in the LDAP drivers

Implements bp list-limiting

Change-Id: I7ca76a8da4260242e578c44103b26257f7e2a5d5