Add netaddr to test requirements
Because of the way we do old keystoneclient tests we can end up with the
situation where an old keystone uses a newer keystoneclient without
doing an install and so misses out on the newer dependencies.

This patch is not a solution to the underlying problem, but should
prevent current gate failings.

Fixes: bug 1212939
Change-Id: I25eca24c6820611a5937a1f709ba42620b70b304