Move CA key from certs directory to private directory
Unlike certs which are public keys are private. The CA key file was
improperly located in the certs directory with the public certs. It's
vital to protect the CA key, it's one of the most security sensitive
files in the system. We already had a private directory for keys,
/etc/keystone/ssl/private where the signing key is located. This fix
moves the CA key into /etc/keystone/ssl/private from
/etc/keystone/ssl/certs.

We also update all relevant files to note the change. During so it was
observered etc/keystone.conf.sample used the wrong parent directory
/etc/keystone/pki, the directory /etc/keystone/ssl is the directory in
actual use.

Fixes: bug #1206254

Change-Id: I014e9f79093a6aa59cd5b3bb6cefa4c6dced67a6