Fixes for user response with LDAP user_enabled_mask
This fixes a couple of problems with the returned user object
when the Keystone server when the LDAP Identity backend is
configured and the user_enabled_mask setting is not zero.

First, the returned user had the masked value (e.g., 512) when
a user was created rather than the actual enabled value (true
or false).

Second, the Keystone server was returning the enabled_nomask
attribute in the user entry when users were listed or updated.

This change makes it so that the server returns the enabled
value as a Boolean when a user is created. Also, the
enabled_nomask attribute is not returned when users are
listed or updated.

Change-Id: I8a22aec3377ffd335095e645c1b3d65008dd0cf2
Closes-bug: #1220945