Monkey patch select in environment
Explicitly say what needs to be monkey patched in
environment/__init__.py. By doing this we are making the code easier to
read and select needs to be monkey patched in order for notifications to
work with keystone/openstack/common/rpc/impl_qpid.py. Without monkey
patching select, keystone will hang on sending a notification since
qpid_messaging won't operate properly when establishing a connection
with the qpid broker.

Change-Id: I4cbdf29924473a76ce459d969e5e7077aac5beb8
Closes-bug: #1224565