Fix live LDAP tests
The live LDAP tests were failing. The tests for case
sensitivity for project name and user name failed because the
attribute is not case sensitive. The test for updating a
missing attribute with a falsy value was failing because the
attribute chosen for the project description does not allow an
empty string.

Change-Id: I723c40b5eab24ac13e27fd040d0d881998f106db
Related-Bug: #1231772