Fixes broken doc references
The references in configuringservices.rst did not resolve to
html addresses and were not referenced correctly in the browser.
This new fix will use the preferred way to cross reference documents.
Has the disadvantage that if the reference changes, it need to be
updated at severeal places in the document.

Change-Id: I3b8f71a95e66266b569aa1141985e47f1f4554c3
Closes-Bug: 1232878