Try decoding string to UTF-8 on error message fail
Database Errors from SQLAlchemy tend to get mangled into ASCII. If an
error is process that contains UTF-8 data in ASCII form if fails to
build the message and will lead to a 5xx error. Catch the error and try
to decode it to create the message again, if that doesn't work at least
fail gracefully.

Change-Id: Iecb1170387c51064918780dc6de07db7ca8aeeee
Closes-Bug: 1253905