PasteConfigNotFound also raised when keystone.conf not found
Although keystone is looking for paste configuration on startup, this
exception can also be raised when keystone attempts to look in the
regular keystone.conf file for paste configuration, and can't find that
file either. The solution here is just to make the exception slightly
more generic (ConfigFileNotFound), along with it's rendered error
message and improve surrounding comments to better explain the scenario.

Change-Id: I88f1d987113884097b1cdf8c120613613abfb8c9
Closes-Bug: 1254036